First stage refactor of GLL parts
