+++ src/main/java/org/nineml/coffeegrinder/gll/MBsrAdd.java
@@ -10,9 +10,6 @@ public class MBsrAdd extends MStatement {
     }
 
     protected void execute(GllParser gllParser) {
-        if (gllParser.trace) {
-            System.err.printf("bsrAdd(%s, %d, %d, %d)\n", slot, gllParser.c_U, gllParser.c_I, gllParser.c_I+1);
-        }
         gllParser.bsrAdd(slot, gllParser.c_U, gllParser.c_I, gllParser.c_I+1);
     }
