+++ src/main/java/org/nineml/coffeegrinder/gll/MIncrementCI.java
@@ -2,11 +2,7 @@ package org.nineml.coffeegrinder.gll;
 
 public class MIncrementCI extends MStatement {
     protected void execute(GllParser gllParser) {
-        gllParser.c_I++;
-        if (gllParser.trace) {
-            System.err.println("c_I = " + gllParser.c_I);
-        }
-
+        gllParser.incrementC_I();
     }
 
     @Override
