+++ src/test/java/org/nineml/coffeegrinder/IxmlTest.java
@@ -18,6 +18,7 @@ public class IxmlTest {
     public void testRuleParse() {
         try {
             ParserOptions options = new ParserOptions();
+            options.getLogger().setDefaultLogLevel(99);
             Grammar grammar = new GrammarCompiler().parse(new File("src/test/resources/ixml.cxml"));
             grammar.setParserOptions(options);
